requirements.txt
aiohttp==3.9.5
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
sqlalchemy==2.0.44
